Convert int to size_t when comparing

+++ src/nt-parser/oracle.cc
@@ -102,7 +102,7 @@ void TopDownOracle::load_oracle(const string& file, bool is_training) {
       }
     }
     actions.push_back(cur_acts);
-    if (termc != sents.back().size()) {
+    if ((size_t) termc != sents.back().size()) {
       cerr << "Mismatched number of tokens and SHIFTs in oracle before line " << lc << endl;
       abort();
     }
@@ -162,7 +162,7 @@ void TopDownOracleGen::load_oracle(const string& file) {
       }
     }
     actions.push_back(cur_acts);
-    if (termc != sents.back().size()) {
+    if ((size_t) termc != sents.back().size()) {
       cerr << "Mismatched number of tokens and SHIFTs in oracle before line " << lc << endl;
       abort();
     }
